MDmisc/requirements.txt
GitPython
psycopg2-binary